--- study_ast/__init__.py
@@ -1,3 +1,0 @@
-# Init file for the module
-def hi():
-    print(1)
